CLI-1362 Refactor manual commit checkpoint IO into submodules
